services/tail_suricata/requirements.txt
requests==2.32.3
watchdog==4.0.1
python-dateutil==2.9.0.post0
